update minor fixng

+++ database/migrations/2020_06_16_154004_create_bobot_aktivitas_table.php
@@ -21,7 +21,7 @@ class CreateBobotAktivitasTable extends Migration
             $table->foreign('ACTIVITIES_ID')->references('id')->on('activites')->onDelete('cascade')->onUpdate('cascade');
             $table->integer('ACADEMIC_YEAR_ID')->unsigned();
             $table->foreign('ACADEMIC_YEAR_ID')->references('id')->on('academic_years')->onDelete('cascade')->onUpdate('cascade');
-            $table->double('WEIGHT');
+            $table->integer('WEIGHT');
             $table->timestamps();
         });
     }
